5-5팔로잉/QLRM lib  DTO Mapping

+++ src/main/java/com/cos/photogramstart/web/dto/subscribe/SubscribeDto.java
@@ -1,5 +1,7 @@
 package com.cos.photogramstart.web.dto.subscribe;
 
+import java.math.BigInteger;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,9 +16,7 @@ public class SubscribeDto {
 	private int Id;
 	private String username;
 	private String profileImageUrl;
-	private Integer subscribeState;
-	private Integer equalUserState;
-	
-	
+	private BigInteger subscribeState;
+	private BigInteger equalUserState;
 	
 }
